+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Support for GoodData.connect ENV['GD_GEM_USER'], ENV['GD_GEM_PASSWORD']
 - Added Schedule#execute(:wait => true|false) option
 - Merge GoodData::Rest::Connection and GoodData::Rest::Connection::RestClientConnection
+- Unified expection handler for REST API and WebDav Access (using GoodData::Rest::Connection.retryable)
 
 ## 0.6.11